requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Model and training
huggingface-hub>=0.15.0
wandb>=0.15.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation and metrics
openai>=0.28.0
anthropic>=0.5.0
evaluate>=0.4.0
rouge-score>=0.1.2
bleurt>=0.0.1

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0

# Data processing and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0
click>=8.1.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development and linting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional: For domain availability checking
dnspython>=2.4.0
whois>=0.9.27

# Optional: For advanced text processing
nltk>=3.8.0
spacy>=3.6.0
textstat>=0.7.3

# Optional: For deployment
gunicorn>=21.2.0
docker>=6.1.0 